+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - nostr: add `Nip11GetOptions` (https://github.com/rust-nostr/nostr/pull/913)
 - nostr: add `RelayUrl::domain` method (https://github.com/rust-nostr/nostr/pull/914)
 - pool: allow putting relays to sleep when idle (https://github.com/rust-nostr/nostr/pull/926)
+- mls: add `NostrMls::add_members` method for adding members to an existing group (https://github.com/rust-nostr/nostr/pull/931)
 
 ### Fixed